First pass of converting filtered HTML to markdown
This doesn't quite do the job because it removes line breaks in the
source content that would otherwise be important in the destination.

I guess I could have left in the <p> and <br> elements in the source by
that was not consistent and would have meant more cleanup.

Maybe I'll try it line by line next